streamlit_app/requirements.txt
numpy
streamlit
tensorflow
pillow